Add FrameDetector bindings for webcam demo
