+++ python-sdk-samples/requirements.txt
@@ -1,4 +1,5 @@
 # Packages to install:
 
-opencv-contrib-python
 tqdm
+numpy
+opencv-contrib-python==4.4.0.42
